More Coins
isMessageSignatureValid now requires a coin parameter to check the address against and use the right magic, due to overlaps in Version Prefix values.
Support for alternate base58 formats such as Ripple included.
Error response for unsupported coins included.

Developer's Note: I'm not sure if the null byte padding between lines 235 and 240 of verifymessage.php is correct, but I haven't run into any addresses that fail because of it, yet. If anyone runs into trouble, please open an Issue or make a Pull Request!